Fix playermonster spellcasting
Prior to this change, 'psibolt' and 'open wounds' were the only spells
available if the playermonster did not have a set spell list (such as for a
minotaur priestess). This change is an overall nerf against enemies that
are immune to touch of death and turn to stone, ie the nastiest ones.

INSECTS is not a player-castable spell, similar to SUMMON_MONS, SUMMON_ANGEL,
etc. Minor note: this behaviour can cause you to uselessly cast CURE_SELF
while at full health.

Adtypes that are not AD_MAGC and AD_CLRC return the message
"You have no spells to cast right now!" when using the (M)onster command,
instead of crashing (because they require a target).

uspell_would_be_useless no longer crashes when given a null target,
like when using the (M)onster command.

Minor buff to player-casted TURN_TO_STONE.